requirements.txt
click==7.1.2
GitPython==3.1.14
PGPy==0.5.3
setuptools==54.1.1
cryptography==3.4.6